+++ src/main/java/com/capstone/ecomm_product/Services/ProductServiceImpl.java
@@ -4,15 +4,16 @@ package com.capstone.ecomm_product.Services;
 import com.capstone.ecomm_product.DTOs.ProductListResponse;
 import com.capstone.ecomm_product.DTOs.ProductRequestDTO;
 import com.capstone.ecomm_product.DTOs.ProductResponseDTO;
-import com.capstone.ecomm_product.Exception.ProductNotFoundException;
+import com.capstone.ecomm_product.Exception.BadRequestClient;
+import com.capstone.ecomm_product.Exception.ResourceNotFoundException;
 import com.capstone.ecomm_product.Models.Category;
-import com.capstone.ecomm_product.Models.Order;
 import com.capstone.ecomm_product.Models.Price;
 import com.capstone.ecomm_product.Models.Product;
 import com.capstone.ecomm_product.Repositories.CategoryRepository;
 import com.capstone.ecomm_product.Repositories.OrderRepository;
 import com.capstone.ecomm_product.Repositories.PriceRepository;
 import com.capstone.ecomm_product.Repositories.ProductRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -21,6 +22,7 @@ import static com.capstone.ecomm_product.Mapper.ProductMapper.convertProductList
 import static com.capstone.ecomm_product.Mapper.ProductMapper.convertProductToProductResponse;
 
 @Service
+@Slf4j
 public class ProductServiceImpl implements ProductService{
 
 
@@ -39,69 +41,6 @@ public class ProductServiceImpl implements ProductService{
         this.prodr = prodr;
     }
 
-    public void initiate() throws ProductNotFoundException {
-        Scanner sc = new Scanner(System.in);
-        System.out.println("Enter Category Name");
-        String categoryName=sc.nextLine();
-        Optional<Category>categoryOptional=cr.findByCategoryName(categoryName);
-        Category c;
-        if(categoryOptional.isEmpty()) {
-             c = new Category();
-            c.setCategoryName(categoryName);
-            c = cr.save(c);
-        }
-        else {
-             c =categoryOptional.get();
-        }
-        while(true) {
-
-            Product p = new Product();
-            p.setCategory(c);
-            p.setImage("http://someImage.com");
-            System.out.println("Enter the product Name for category " + categoryName);
-            String title = sc.nextLine();
-            p.setTitle(title);
-            System.out.println("Enter the Description of Product " + title);
-            String description=sc.nextLine();
-            p.setDescription(description);
-            System.out.println("Enter the price of product " + title);
-            Price p1 = new Price();
-            p1.setCurrency("INR");
-            double amount = sc.nextDouble();
-            p1.setAmount(amount);
-            p1.setDiscount(0.00);
-            p1=pr.save(p1);
-            p.setPrice(p1);
-            sc.nextLine();
-            System.out.println("DO you want to exit ?");
-            String val=sc.nextLine();
-            p=prodr.save(p);
-            if(val.equalsIgnoreCase("yes"))
-                break;
-
-        }
-        Order order = new Order();
-        List<Product> productList= new ArrayList<>();
-        while (true){
-            System.out.println("Enetr the products which you want to add in cart? ");
-            String title =sc.nextLine();
-            Optional<Product> productOptional = prodr.findByTitle(title);
-            if (productOptional.isEmpty()){
-                throw new ProductNotFoundException();
-            }
-            productList.add(productOptional.get());
-            System.out.println("Do you want to Exit Card?");
-            String val=sc.nextLine();
-            if(val.equalsIgnoreCase("yes")) {
-                order.setProductList(productList);
-                break;
-            }
-        }
-        or.save(order);
-
-
-
-    }
 
     @Override
     public ProductListResponse getAllProducts() {
@@ -111,35 +50,37 @@ public class ProductServiceImpl implements ProductService{
     }
 
     @Override
-    public ProductResponseDTO getProductById(UUID id) throws ProductNotFoundException {
-
+    public ProductResponseDTO getProductById(UUID id) throws ResourceNotFoundException {
+        log.info("Service finding product with product id {} from DB",id);
         Optional<Product> productOptional=prodr.findById(id);
         if(productOptional.isEmpty()){
-            throw new ProductNotFoundException();
+            log.error("Product ID with {} does not exists in DB",id);
+            throw new ResourceNotFoundException("Product","Product ID",id);
         }
         ProductResponseDTO responseDTO= convertProductToProductResponse(productOptional.get());
-
+        log.info("Product with product id {} found",id);
         return responseDTO;
     }
 
     @Override
-    public ProductResponseDTO createProduct(ProductRequestDTO requestDTO) {
+    public ProductResponseDTO createProduct(ProductRequestDTO requestDTO) throws ResourceNotFoundException {
         Product p = new Product();
+        log.info("Starting product creation for product price: {}", requestDTO.getPrice());
         Price p1 = new Price();
         p1.setAmount(requestDTO.getPrice());
         p1.setDiscount(0.00);
         p1.setCurrency(requestDTO.getCurrency());
+        log.info("Saving price details: amount={}, currency=INR", requestDTO.getPrice());
         p1=pr.save(p1);
         p.setPrice(p1);
         Category c;
         Optional<Category> categoryOptional =cr.findByCategoryName(requestDTO.getCategory());
         if(categoryOptional.isEmpty()) {
-             c = new Category();
-            c.setCategoryName(requestDTO.getCategory());
-            c.setCategoryAddedAt(new Date());
-            c=cr.save(c);
+            log.error("Category not found: {}", requestDTO.getCategory());
+            throw new ResourceNotFoundException("Category","CategoryName",requestDTO.getCategory());
         }
         else {
+            log.info("Category found: {}", requestDTO.getCategory());
             c=categoryOptional.get();
         }
         p.setCategory(c);
@@ -147,23 +88,50 @@ public class ProductServiceImpl implements ProductService{
         p.setDescription(requestDTO.getDescription());
         p.setTitle(requestDTO.getTitle());
         p.setProductAddedAt(new Date());
+        log.info("Saving product: title={}, description={}", p.getTitle(), p.getDescription());
         p=prodr.save(p);
-
+        log.info("Product created successfully with ID: {}", p.getId());
 return convertProductToProductResponse(p);
     }
 
     @Override
-    public boolean deleteProduct(UUID id) throws ProductNotFoundException {
-        try {
-            prodr.deleteById(id);
-        } catch (Exception e){
-            throw new ProductNotFoundException();
+    public boolean deleteProduct(UUID id) throws ResourceNotFoundException {
+        Optional<Product> productOptional = prodr.findById(id);
+        if(productOptional.isEmpty()){
+            log.error("Product Not Found with id {}",id);
+            throw new ResourceNotFoundException("Delete Product", "Product Id",id);
+        }
+        else{
+           prodr.deleteById(id);
+           log.info("Product with product id {} deleted successfully",id);
         }
         return true;
     }
 
     @Override
-    public Product updateProduct(int id, Product updateProduct) {
-        return null;
+    public ProductResponseDTO updateProduct(UUID id, ProductRequestDTO updateProduct) throws BadRequestClient {
+        Optional<Product> productOptional=prodr.findById(id);
+        if(productOptional.isEmpty()){
+            log.error("Product Not Found with id {}",id);
+            throw new ResourceNotFoundException("Product","Product ID",id);
+        }
+        else if(!productOptional.get().getCategory().getCategoryName().equals(updateProduct.getCategory()))
+        {
+            log.error("Category not found: {}", updateProduct.getCategory());
+            throw new BadRequestClient("Category", "CategoryName",updateProduct.getCategory());
+        }
+        else {
+          Product product= productOptional.get();
+          product.setTitle(updateProduct.getTitle());
+          product.setImage(updateProduct.getImage());
+          product.setDescription(updateProduct.getDescription());
+          Price p = new Price();
+          p.setAmount(updateProduct.getPrice());
+          product.setPrice(p);
+            log.info("Product with product id {} edited successfully",id);
+            return convertProductToProductResponse(product);
+        }
     }
+
+
 }
